Extract browser visual compare runner
